+++ src/gufo/thor/services/bi.py
@@ -18,13 +18,14 @@ from .login import login
 from .migrate import migrate
 from .noc import NocService
 from .static import static
+from .web import web
 
 
 class BiService(NocService):
     """bi service."""
 
     name = "bi"
-    dependencies = (auth, clickhouse, envoy, login, migrate, static)
+    dependencies = (auth, clickhouse, envoy, login, migrate, static, web)
     expose_http_prefix = "/api/bi/"