+++ src/gufo/thor/services/card.py
@@ -20,6 +20,7 @@ from .mongo import mongo
 from .noc import NocService
 from .postgres import postgres
 from .static import static
+from .web import web
 
 
 class CardService(NocService):
@@ -35,6 +36,7 @@ class CardService(NocService):
         mongo,
         postgres,
         static,
+        web,
     )
     expose_http_prefix = "/api/card/"
     require_http_auth = True